refactor: format 관련 로직 파일명 수정
