+++ client/src/entities/stopwatch/ui/formatters.ts
