+++ client/src/features/timer/index.ts
@@ -1,4 +1,4 @@
 export * from "./model/types";
 export * from "./model/reducers";
 export * from "./model/selectors";
-export { formatMsToTime } from "./ui/format-time";
+export { formatMsToTime } from "./ui/formatters";
